Handle broken pipe gracefully
Previously, the code would panic if it couldn't write
to the supplied buffer. Now, the code will check why it
can't write to the supplied buffer. If the reason is a
broken pipe, i.e., the file it's writing to is closed,
it will exit gracefully. Otherwise, it will surface the error
and exit 1.